requirements.txt
fastapi
uvicorn
openai
python-dotenv
duckdb
pandas
numpy
faiss-cpu